requirements.txt
# Gunicorn
gunicorn

# Django Auth
django-cors-headers
python-decouple

# Django Image Processing
pillow

# Supabase Integration(postgres)
psycopg2
